Extract slide-in-text
